feat(dispatcher): only delete np msg if not 1st

+++ src/util/dispatcher.js
@@ -29,6 +29,15 @@ export default class Dispatcher {
                 else if (this.repeat === 'all' || this.repeat === 'off') {
                     _notifiedOnce = false;
                 }
+                if (this.nowPlayingMessage) {
+                    const msgs = await this.channel.messages.fetch({ limit: 1 });
+                    if (msgs.filter(m => m.id === this.nowPlayingMessage.id)) {
+                        await this.nowPlayingMessage
+                            .edit({ embeds: [ container.client.util.embed(`${container.client.config.emojis.playing} [**${this.current.info.title}** - **${this.current.info.author}**](${this.current.info.uri}) \`${Dispatcher.humanizeTime(this.current.info.length)}\` (${this.current.info.requester.toString()})`) ] })
+                            .catch(() => null);
+                        return;
+                    }
+                }
                 this.nowPlayingMessage = await this.channel
                     .send({ embeds: [ container.client.util.embed(`${container.client.config.emojis.playing} [**${this.current.info.title}** - **${this.current.info.author}**](${this.current.info.uri}) \`${Dispatcher.humanizeTime(this.current.info.length)}\` (${this.current.info.requester.toString()})`) ] })
                     .catch(() => null);
@@ -37,6 +46,8 @@ export default class Dispatcher {
                 if (this.repeat === 'one') this.queue.unshift(this.current);
                 if (this.repeat === 'all' && !this.current.skipped) this.queue.push(this.current);
                 if (this.nowPlayingMessage && this.repeat !== 'one') {
+                    const msgs = await this.channel.messages.fetch({ limit: 1 });
+                    if (msgs.filter(m => m.id === this.nowPlayingMessage.id)) return;
                     await this.nowPlayingMessage.delete().catch(() => null);
                     this.nowPlayingMessage = null;
                 }
